api/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
scikit-learn==0.24.2
python-whois
bs4
gradio_client
requests